requirements.txt
fastapi
uvicorn
motor
pymongo
numpy
pandas
scipy.sparse.linalg